requirements/prod.txt
asgiref==3.4.1
astroid==2.6.2
colorama==0.4.4
Django==2.2.24
djangorestframework==3.12.4
gunicorn==20.1.0
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.9.3
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
wrapt==1.12.1
